calender_app/requirements.txt
streamlit
python-dotenv
google-generativeai
streamlit-calendar
streamlit-modal